Ready for first pass at CF

+++ ras_common/ons_database.py
@@ -26,12 +26,21 @@ class ONSDatabase(object):
         self._base = declarative_base()
         self._session = scoped_session(sessionmaker())
 
+    def log(self, text):
+        self._env.logger.info('[db] {}'.format(text))
+
+    def warn(self, text):
+        self._env.logger.warn('[db] [warning] {}'.format(text))
+
     def activate(self):
         if self._env.get('enable_database', 'false').lower() not in ['true', 'yes']:
-            return self._env.logger.info('Database is NOT enabled [missing "enabled_database = true"]')
+            return self.log('Database is NOT enabled [missing "enabled_database = true"]')
+
+        if not self.check_paths():
+            return self.warn('[swagger_server/models_local/_models.py] file is missing')
 
         db_connection = self._env.get('db_connection')
-        self._env.logger.info('Database connection is "{}"'.format(db_connection))
+        self.log('Database connection is "{}"'.format(db_connection))
         self._engine = create_engine(db_connection, convert_unicode=True)
         self._session.remove()
         if self._env.drop_database:
@@ -43,39 +52,31 @@ class ONSDatabase(object):
         Check our filesystem for required database files ...
         :return: True if database environment is available
         """
-        if not Path('swagger_server').is_dir():
-            return self._env.logger.warn('[swagger_server] folder is missing')
-        if not Path('swagger_server/models_local').is_dir():
-            return self._env.logger.warn('[swagger_server/models_local] folder is missing')
-        if not Path('swagger_server/models_local/_models.py').is_file():
-            return self._env.logger.warn('[swagger_server/models_local/_models.py] file is missing')
+        return Path('swagger_server/models_local/_models.py').is_file()
 
     def drop(self):
-        self._env.logger.info('Dropping database tables')
-        if not self.check_paths():
-            return
+        self.log('Dropping any existing Database Tables')
         from swagger_server.models_local import _models
         connection = self._env.get('db_connection')
         schema = self._env.get('db_schema')
         if connection.startswith('postgres'):
-            self._env.logger.info('Dropping schema "{}"'.format(schema))
+            self.log('Dropping pre-existing schema "{}" if it exists'.format(schema))
             self._base.metadata.schema = schema
             self._engine.execute("DROP SCHEMA IF EXISTS {} CASCADE".format(schema))
         else:
             self._base.metadata.drop_all(self._engine)
 
     def create(self):
-        self._env.logger.info('Creating database tables')
+        self.log('Creating any missing Database tables')
         from swagger_server.models_local import _models
         connection = self._env.get('db_connection')
         schema = self._env.get('db_schema')
         if connection.startswith('postgres'):
-            self._env.logger.info('Creating schema "{}"'.format(schema))
+            self.log('Creating Database schema "{}"'.format(schema))
             self._base.metadata.schema = schema
 
-        self._env.logger.info('Creating database with uri "{}"'.format(connection))
         if connection.startswith('postgres'):
-            self._env.logger.info("Creating schema {}.".format(schema))
+            self.log("Creating schema {} if it does't exist".format(schema))
             self._engine.execute("CREATE SCHEMA IF NOT EXISTS {}".format(schema))
-        self._env.logger.info("Creating database tables.")
+        self.log("Running Create-All")
         self._base.metadata.create_all(self._engine)
